requirements.txt
Django == 3.1.6
Djangorestframework == 3.12.2